requirements.txt
Pillow==10.0.0
pika==1.3.1
redis==4.5.4
boto3==1.26.137
python-dotenv==1.0.0